[RENAME]: FriendListResponseModel Share>Model로 이동, Info -> FriendDefaultInfo로 변경
