+++ NEMODU/NEMODU/Screen/Share/Model/FriendsListResponseModel.swift
@@ -1,0 +1,15 @@
+//
+//  FriendsListResponseModel.swift
+//  NEMODU
+//
+//  Created by Kim HeeJae on 2022/11/01.
+//
+
+import Foundation
+import Alamofire
+
+struct FriendsListResponseModel: Codable {
+    let infos: [FriendDefaultInfo]
+    let isLast: Bool
+    let size: Int
+}
